Moved view model logic to a service
