+++ app/uk/gov/hmrc/tai/viewModels/taxCodeChange/TaxCodeChangeViewModel.scala
@@ -27,82 +27,17 @@ import uk.gov.hmrc.tai.util.constants.TaiConstants
 case class TaxCodeChangeViewModel(pairs: TaxCodePairs,
                                   changeDate: LocalDate,
                                   scottishTaxRateBands: Map[String, BigDecimal],
-                                  gaDimensions: Map[String, String]) {
-
-  private def removeEmployer(employerNames: Seq[String])(implicit messages: Messages): Seq[String] = {
-    employerNames map (name => messages("tai.taxCodeComparison.removeEmployer", name))
-  }
-
-  private def addEmployer(employerNames: Seq[String])(implicit messages: Messages): Seq[String] = {
-    employerNames map (name => messages("tai.taxCodeComparison.addEmployer", name))
-  }
-
-  private def genericMessage(implicit messages: Messages): Seq[String] = {
-    Seq(messages("taxCode.change.yourTaxCodeChanged.paragraph"))
-  }
-
-  private def secondaryEmploymentsChanged(implicit messages: Messages): Seq[String] = {
-    val previous = pairs.unMatchedPreviousCodes.flatMap(_.previous).map(record => record.employerName)
-    val current = pairs.unMatchedCurrentCodes.flatMap(_.current).map(record => record.employerName)
-
-    val uniquePrevious = previous.distinct.sorted
-    val uniqueCurrent = current.distinct.sorted
-
-    val currentAndPreviousEmployerNamesAreSame: Boolean = (uniquePrevious == uniqueCurrent) && (uniquePrevious ++ uniqueCurrent).nonEmpty
-
-    currentAndPreviousEmployerNamesAreSame match {
-      case true => genericMessage
-      case false => removeEmployer(previous) ++ addEmployer(current)
-    }
-  }
-
-  private def isDifferentPayRollWithSameEmployerName(primaryPair: TaxCodePair): Boolean = {
-
-    val isEmployerNameSame = {
-      val currentName = primaryPair.current.map(_.employerName)
-      val previousName = primaryPair.previous.map(_.employerName)
-
-      currentName == previousName
-    }
-
-    val isPayRollSame = {
-      val previousPayRoll = primaryPair.previous.flatMap(_.payrollNumber)
-      val currentPayRoll = primaryPair.current.flatMap(_.payrollNumber)
-
-      previousPayRoll == currentPayRoll
-    }
-
-    isEmployerNameSame && !isPayRollSame
-  }
-
-  private def primaryEmploymentsChanged(implicit messages: Messages): Seq[String] = {
-    pairs.primaryPairs flatMap { primaryPair: TaxCodePair =>
-      val current = primaryPair.current.map(_.employerName)
-      val previous = primaryPair.previous.map(_.employerName)
-
-      (current, previous) match {
-        case (Some(current), Some(previous)) if (current != previous) => removeEmployer(Seq(previous)) ++ addEmployer(Seq(current))
-        case (Some(current), Some(previous)) if isDifferentPayRollWithSameEmployerName(primaryPair) => {
-          genericMessage
-        }
-        case _ => Seq.empty[String]
-      }
-    }
-  }
-
-  def taxCodeReasons(implicit messages: Messages): Seq[String] = {
-    primaryEmploymentsChanged ++ secondaryEmploymentsChanged
-  }
-}
+                                  taxCodeChangeReasons: Seq[String],
+                                  gaDimensions: Map[String, String])
 
 object TaxCodeChangeViewModel extends TaxCodeDescriptor {
 
-  def apply(taxCodeChange: TaxCodeChange, scottishTaxRateBands: Map[String, BigDecimal]): TaxCodeChangeViewModel = {
+  def apply(taxCodeChange: TaxCodeChange, scottishTaxRateBands: Map[String, BigDecimal], taxCodeChangeReasons: Seq[String] = Seq.empty[String]): TaxCodeChangeViewModel = {
 
-    val taxCodePairs = TaxCodePairs(taxCodeChange.previous, taxCodeChange.current)
+    val taxCodePairs = TaxCodePairs(taxCodeChange)
     val changeDate = taxCodeChange.mostRecentTaxCodeChangeDate
 
-    TaxCodeChangeViewModel(taxCodePairs, changeDate, scottishTaxRateBands, gaDimensions(taxCodeChange, changeDate))
+    TaxCodeChangeViewModel(taxCodePairs, changeDate, scottishTaxRateBands, taxCodeChangeReasons, gaDimensions(taxCodeChange, changeDate))
   }
 
   def getTaxCodeExplanations(taxCodeRecord: TaxCodeRecord, scottishTaxRateBands: Map[String, BigDecimal], identifier: String)
