+++ app/views/taxCodeChange/taxCodeComparison.scala.html
@@ -49,7 +49,7 @@
 
     <div class="column-two-thirds">
         <div class="grid-row">
-            @displayTaxCodeReasons(taxCodeChangeViewModel.taxCodeReasons)
+            @displayTaxCodeReasons(taxCodeChangeViewModel.taxCodeChangeReasons)
         </div>
         <div class="grid-row tax-code-change margin-bottom-xlarge">
             @taxCodePairs(taxCodeChangeViewModel.pairs)
