+++ test/uk/gov/hmrc/tai/viewModels/taxCodeChange/TaxCodeChangeViewModelSpec.scala
@@ -196,116 +196,4 @@ class TaxCodeChangeViewModelSpec extends PlaySpec with FakeTaiPlayApplication {
       }
     }
   }
-
-  "TaxCodeChangeViewModel taxCodeReasons" must {
-
-    val previousEmployer = "Previous Employer"
-    val currentEmployer = "Current Employer"
-
-    def createTaxRecord(employerName: String): TaxCodeRecord = {
-      TaxCodeRecord("taxCode", startDate, startDate.plusMonths(1), OtherBasisOfOperation, employerName, false, Some("12345"), false)
-    }
-
-    def createPrimaryTaxRecord(employerName: String): TaxCodeRecord = {
-      TaxCodeRecord("taxCode", startDate, startDate.plusMonths(1), OtherBasisOfOperation, employerName, false, Some("12345"), true)
-    }
-
-    def removedEmployer(employerName: String): String = {
-      Messages("tai.taxCodeComparison.removeEmployer", employerName)
-    }
-
-    def addedEmployer(employerName: String): String = {
-      Messages("tai.taxCodeComparison.addEmployer", employerName)
-    }
-
-    "return empty when nothing has changed" in {
-      val taxRecord = Seq(createTaxRecord(previousEmployer), createPrimaryTaxRecord(currentEmployer))
-
-      val taxCodeChange = TaxCodeChange(taxRecord, taxRecord)
-      val model = TaxCodeChangeViewModel(taxCodeChange, Map.empty[String, BigDecimal])
-
-      model.taxCodeReasons mustBe Seq.empty[String]
-    }
-
-    "return a reason when an employment been removed" in {
-      val previous = Seq(createTaxRecord(previousEmployer), createTaxRecord(currentEmployer))
-      val current = Seq(createTaxRecord(currentEmployer))
-
-      val taxCodeChange = TaxCodeChange(previous, current)
-      val model = TaxCodeChangeViewModel(taxCodeChange, Map.empty[String, BigDecimal])
-
-      model.taxCodeReasons mustBe Seq(removedEmployer(previousEmployer))
-    }
-
-    "return a reason when an employment been added" in {
-      val previous = Seq(createTaxRecord(previousEmployer))
-      val current = Seq(createTaxRecord(previousEmployer), createTaxRecord(currentEmployer))
-
-      val taxCodeChange = TaxCodeChange(previous, current)
-      val model = TaxCodeChangeViewModel(taxCodeChange, Map.empty[String, BigDecimal])
-
-      model.taxCodeReasons mustBe Seq(addedEmployer(currentEmployer))
-    }
-
-    "return multiple reasons when employments have changed" in {
-      val previous = Seq(createTaxRecord(previousEmployer), createTaxRecord(previousEmployer + "1"))
-      val current = Seq(createTaxRecord(currentEmployer), createTaxRecord(currentEmployer + "1"))
-
-      val taxCodeChange = TaxCodeChange(previous, current)
-      val model = TaxCodeChangeViewModel(taxCodeChange, Map.empty[String, BigDecimal])
-
-      model.taxCodeReasons mustBe Seq(
-        removedEmployer(previousEmployer), removedEmployer(previousEmployer + "1"),
-        addedEmployer(currentEmployer), addedEmployer(currentEmployer + "1")
-      )
-    }
-
-      "return primary and secondary employment have changed at the same time" in {
-        val previous = Seq(createPrimaryTaxRecord(previousEmployer), createTaxRecord(previousEmployer + "1"))
-        val current = Seq(createPrimaryTaxRecord(currentEmployer), createTaxRecord(currentEmployer + "1"))
-
-        val taxCodeChange = TaxCodeChange(previous, current)
-        val model = TaxCodeChangeViewModel(taxCodeChange, Map.empty[String, BigDecimal])
-
-        model.taxCodeReasons mustBe Seq(
-          removedEmployer(previousEmployer), addedEmployer(currentEmployer),
-          removedEmployer(previousEmployer + "1"), addedEmployer(currentEmployer + "1")
-        )
-    }
-
-    "return an add and remove message when primary employment has changed" in {
-      val previous = Seq(createPrimaryTaxRecord(previousEmployer))
-      val current = Seq(createPrimaryTaxRecord(currentEmployer))
-
-      val taxCodeChange = TaxCodeChange(previous, current)
-      val model = TaxCodeChangeViewModel(taxCodeChange, Map.empty[String, BigDecimal])
-
-      model.taxCodeReasons mustBe Seq(
-        removedEmployer(previousEmployer),
-        addedEmployer(currentEmployer)
-      )
-    }
-
-    "if you can match by employer name but can't match with payroll" should {
-      "go to the generic message for primary tax records" in {
-        val previous = createPrimaryTaxRecord(previousEmployer)
-        val current = previous.copy(payrollNumber = Some("54321"))
-
-        val taxCodeChange = TaxCodeChange(Seq(previous), Seq(current))
-        val model = TaxCodeChangeViewModel(taxCodeChange, Map.empty[String, BigDecimal])
-
-        model.taxCodeReasons mustBe Seq(Messages("taxCode.change.yourTaxCodeChanged.paragraph"))
-      }
-
-      "go to the generic message for secondary tax records" in {
-        val previous = createTaxRecord(previousEmployer)
-        val current = previous.copy(payrollNumber = Some("54321"))
-
-        val taxCodeChange = TaxCodeChange(Seq(previous), Seq(current))
-        val model = TaxCodeChangeViewModel(taxCodeChange, Map.empty[String, BigDecimal])
-
-        model.taxCodeReasons mustBe Seq(Messages("taxCode.change.yourTaxCodeChanged.paragraph"))
-      }
-    }
-  }
 }
